+++ CHANGELOG.md
@@ -13,6 +13,7 @@ but this project DOES NOT adhere to [Semantic Versioning](http://semver.org/).
 * Remove LeanCode.ExternalIdentityProviders
 * Remove LeanCode.PdfRocket
 * Remove StyleCop completely
+* JSON serializer for CQRS now shares options with ASP.NET Core's `JsonOptions` by default
 
 ## 8.1